fixing continued

+++ src/MP3Track.cpp
@@ -32,8 +32,8 @@ void MP3Track::load() {
 
 void MP3Track::analyze_beatgrid() {
     std::cout << "[MP3Track:: analyze_beatgrid] Analyzing beat grid for: \"" << title << "\"" << std::endl;
-    double estimated_beats = (duration_seconds / 60.0) * bpm;
-    double precision_factor = bitrate / 320.0;
+    int estimated_beats = (duration_seconds / 60.0) * bpm;
+    int precision_factor = bitrate / 320.0;
     std::cout << "  -> Estimated beats: " << estimated_beats 
               << " | Compression precision factor: " << precision_factor << std::endl;
